+++ wasm-rpc-stubgen/src/model/mod.rs
@@ -1,2 +1,2 @@
-pub mod component;
+pub mod wasm_rpc;
 pub mod oam;
