+++ wasm-rpc-stubgen/src/model/wasm_rpc.rs
